refactor: 유틸 파일명 storage를 localStorage로 수정
- 보관함 페이지명과의 혼란을 방지하기 위해 유틸 파일명 수정
